Squashed commit of the following:
commit 5bd54ee79c95156fff37ce9935e4377a64dca637
Author: Hidde Wieringa <hidde@hiddewieringa.nl>
Date:   Mon Jul 22 21:37:35 2024 +0200

    deploy only API

commit 2e56b7784c223adc781e535064458b8c256897c2
Author: Hidde Wieringa <hidde@hiddewieringa.nl>
Date:   Mon Jul 22 21:37:18 2024 +0200

    skip cleanup

+++ api/prepare-api.sh
@@ -53,5 +53,5 @@ echo 'Building API container'
 docker compose build api
 
 echo 'Cleaning up'
-rm -rf api/postgres-data
-rm api/api.sql
+[ -n "$SKIP_CLEANUP" ] || rm -rf api/postgres-data
+[ -n "$SKIP_CLEANUP" ] || rm -f api/api.sql
